rename file _dsearch.py to _search.py (#1043)
