+++ pyserini/search/faiss/_searcher.py
